Refactor preference package structure
