+++ lawnchair/src/app/lawnchair/ui/preferences/about/ContributorCard.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package app.lawnchair.ui.preferences
+package app.lawnchair.ui.preferences.about
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
