+++ lawnchair/src/app/lawnchair/ui/preferences/about/LawnchairLink.kt
@@ -14,18 +14,16 @@
  * limitations under the License.
  */
 
-package app.lawnchair.ui.preferences
+package app.lawnchair.ui.preferences.about
 
 import android.content.Intent
 import android.net.Uri
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.layout.*
 import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -36,27 +34,29 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 
 @Composable
-fun ContributorLink(@DrawableRes iconResId: Int, url: String) {
+fun LawnchairLink(@DrawableRes iconResId: Int, label: String, modifier: Modifier = Modifier, url: String) {
     val context = LocalContext.current
 
-    Box(
-        modifier = Modifier
-            .width(48.dp)
-            .height(48.dp)
-            .clip(CircleShape)
+    Column(
+        verticalArrangement = Arrangement.Center,
+        modifier = modifier
+            .height(64.dp)
+            .clip(MaterialTheme.shapes.medium)
             .clickable {
                 val webpage = Uri.parse(url)
                 val intent = Intent(Intent.ACTION_VIEW, webpage)
                 if (intent.resolveActivity(context.packageManager) != null) {
                     context.startActivity(intent)
                 }
-            }
+            },
+        horizontalAlignment = Alignment.CenterHorizontally
     ) {
         Image(
             painterResource(id = iconResId), contentDescription = null, modifier = Modifier
                 .height(24.dp)
-                .width(24.dp)
-                .align(Alignment.Center), colorFilter = ColorFilter.tint(color = MaterialTheme.colors.onBackground)
+                .width(24.dp), colorFilter = ColorFilter.tint(color = MaterialTheme.colors.onBackground)
         )
+        Spacer(modifier = Modifier.requiredHeight(4.dp))
+        Text(text = label, style = MaterialTheme.typography.body1, color = MaterialTheme.colors.onBackground)
     }
 }
