+++ lawnchair/src/app/lawnchair/ui/preferences/components/NavigationActionPreference.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package app.lawnchair.ui.preferences
+package app.lawnchair.ui.preferences.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
@@ -26,12 +26,15 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import androidx.navigation.compose.navigate
 
 @Composable
-fun ClickListenerPreference(
+fun NavigationActionPreference(
     label: String,
     subtitle: String? = null,
-    onClick: () -> Unit,
+    navController: NavController,
+    destination: String,
     showDivider: Boolean = true
 ) =
     PreferenceTemplate(height = if (subtitle != null) 72.dp else 52.dp, showDivider = showDivider) {
@@ -40,7 +43,7 @@ fun ClickListenerPreference(
             modifier = Modifier
                 .fillMaxWidth()
                 .fillMaxHeight()
-                .clickable(onClick = onClick)
+                .clickable { navController.navigate(route = destination) }
                 .padding(start = 16.dp, end = 16.dp),
         ) {
             Text(text = label, style = MaterialTheme.typography.subtitle1, color = MaterialTheme.colors.onBackground)
