+++ lawnchair/src/app/lawnchair/ui/preferences/components/PreferenceCategoryListItem.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package app.lawnchair.ui.preferences
+package app.lawnchair.ui.preferences.components
 
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.Image
