+++ lawnchair/src/app/lawnchair/ui/preferences/components/PreferenceGroup.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package app.lawnchair.ui.preferences
+package app.lawnchair.ui.preferences.components
 
 import androidx.compose.animation.*
 import androidx.compose.foundation.layout.*
