+++ lawnchair/src/app/lawnchair/ui/preferences/components/PreferenceTemplate.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package app.lawnchair.ui.preferences
+package app.lawnchair.ui.preferences.components
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.Divider
